Extract workspace project APIs
